Fix: Protect against segfault while printing a time_t variable. Detected in NetBSD6 RC2(32-bit)

+++ src/logging.c
@@ -478,7 +478,7 @@ void PromiseLog(char *s)
         return;
     }
 
-    fprintf(fout, "%ld,%ld: %s\n", CFSTARTTIME, now, s);
+    fprintf(fout, "%jd,%jd: %s\n", (intmax_t) CFSTARTTIME, (intmax_t) now, s);
     fclose(fout);
 }
